Session code (#2)
session code

This commit implements the functionality for managing sessions. The following features have been added:

Create a new session
Show a list of sessions
Show the details of a specific session
Edit the details of a specific session
Delete a specific session
The database schema has been updated to include a sessions table with the necessary columns.